langchain/requirements.txt
langchain
python-dotenv
langchain[groq]